requirements.txt
uvicorn==0.23.2
fastapi==0.104.1
pydantic==2.4.2
motor==3.3.1
python-dotenv==1.0.0
logging==0.4.9.6
pymongo==4.6.0
tornado==6.3.3